[admin] extract stats handlers to separate file (#10750)
