+++ test/server/http/admin_test.cc
@@ -12,27 +12,17 @@
 #include "envoy/extensions/transport_sockets/tls/v3/cert.pb.h"
 #include "envoy/json/json_object.h"
 #include "envoy/runtime/runtime.h"
-#include "envoy/stats/stats.h"
 
 #include "common/http/message_impl.h"
 #include "common/json/json_loader.h"
 #include "common/profiler/profiler.h"
 #include "common/protobuf/protobuf.h"
 #include "common/protobuf/utility.h"
-#include "common/stats/symbol_table_creator.h"
-#include "common/stats/thread_local_store.h"
-
-#include "server/http/admin.h"
-#include "server/http/admin_filter.h"
 
 #include "extensions/transport_sockets/tls/context_config_impl.h"
 
-#include "test/mocks/http/mocks.h"
-#include "test/mocks/runtime/mocks.h"
-#include "test/mocks/server/mocks.h"
-#include "test/test_common/environment.h"
+#include "test/server/http/admin_instance.h"
 #include "test/test_common/logging.h"
-#include "test/test_common/network_utility.h"
 #include "test/test_common/printers.h"
 #include "test/test_common/utility.h"
 
@@ -41,10 +31,8 @@
 #include "gtest/gtest.h"
 
 using testing::AllOf;
-using testing::EndsWith;
 using testing::Ge;
 using testing::HasSubstr;
-using testing::InSequence;
 using testing::Invoke;
 using testing::NiceMock;
 using testing::Property;
@@ -52,552 +40,9 @@ using testing::Ref;
 using testing::Return;
 using testing::ReturnPointee;
 using testing::ReturnRef;
-using testing::StartsWith;
 
 namespace Envoy {
 namespace Server {
-
-class AdminStatsTest : public testing::TestWithParam<Network::Address::IpVersion> {
-public:
-  AdminStatsTest()
-      : symbol_table_(Stats::SymbolTableCreator::makeSymbolTable()), alloc_(*symbol_table_) {
-    store_ = std::make_unique<Stats::ThreadLocalStoreImpl>(alloc_);
-    store_->addSink(sink_);
-  }
-
-  static std::string
-  statsAsJsonHandler(std::map<std::string, uint64_t>& all_stats,
-                     std::map<std::string, std::string>& all_text_readouts,
-                     const std::vector<Stats::ParentHistogramSharedPtr>& all_histograms,
-                     const bool used_only, const absl::optional<std::regex> regex = absl::nullopt) {
-    return AdminImpl::statsAsJson(all_stats, all_text_readouts, all_histograms, used_only, regex,
-                                  true /*pretty_print*/);
-  }
-
-  Stats::SymbolTablePtr symbol_table_;
-  NiceMock<Event::MockDispatcher> main_thread_dispatcher_;
-  NiceMock<ThreadLocal::MockInstance> tls_;
-  Stats::AllocatorImpl alloc_;
-  Stats::MockSink sink_;
-  std::unique_ptr<Stats::ThreadLocalStoreImpl> store_;
-};
-
-INSTANTIATE_TEST_SUITE_P(IpVersions, AdminStatsTest,
-                         testing::ValuesIn(TestEnvironment::getIpVersionsForTest()),
-                         TestUtility::ipTestParamsToString);
-
-TEST_P(AdminStatsTest, StatsAsJson) {
-  InSequence s;
-  store_->initializeThreading(main_thread_dispatcher_, tls_);
-
-  Stats::Histogram& h1 = store_->histogramFromString("h1", Stats::Histogram::Unit::Unspecified);
-  Stats::Histogram& h2 = store_->histogramFromString("h2", Stats::Histogram::Unit::Unspecified);
-
-  EXPECT_CALL(sink_, onHistogramComplete(Ref(h1), 200));
-  h1.recordValue(200);
-
-  EXPECT_CALL(sink_, onHistogramComplete(Ref(h2), 100));
-  h2.recordValue(100);
-
-  store_->mergeHistograms([]() -> void {});
-
-  // Again record a new value in h1 so that it has both interval and cumulative values.
-  // h2 should only have cumulative values.
-  EXPECT_CALL(sink_, onHistogramComplete(Ref(h1), 100));
-  h1.recordValue(100);
-
-  store_->mergeHistograms([]() -> void {});
-
-  std::vector<Stats::ParentHistogramSharedPtr> histograms = store_->histograms();
-  std::sort(histograms.begin(), histograms.end(),
-            [](const Stats::ParentHistogramSharedPtr& a,
-               const Stats::ParentHistogramSharedPtr& b) -> bool { return a->name() < b->name(); });
-  std::map<std::string, uint64_t> all_stats;
-  std::map<std::string, std::string> all_text_readouts;
-  std::string actual_json = statsAsJsonHandler(all_stats, all_text_readouts, histograms, false);
-
-  const std::string expected_json = R"EOF({
-    "stats": [
-        {
-            "histograms": {
-                "supported_quantiles": [
-                    0.0,
-                    25.0,
-                    50.0,
-                    75.0,
-                    90.0,
-                    95.0,
-                    99.0,
-                    99.5,
-                    99.9,
-                    100.0
-                ],
-                "computed_quantiles": [
-                    {
-                        "name": "h1",
-                        "values": [
-                            {
-                                "interval": 100.0,
-                                "cumulative": 100.0
-                            },
-                            {
-                                "interval": 102.5,
-                                "cumulative": 105.0
-                            },
-                            {
-                                "interval": 105.0,
-                                "cumulative": 110.0
-                            },
-                            {
-                                "interval": 107.5,
-                                "cumulative": 205.0
-                            },
-                            {
-                                "interval": 109.0,
-                                "cumulative": 208.0
-                            },
-                            {
-                                "interval": 109.5,
-                                "cumulative": 209.0
-                            },
-                            {
-                                "interval": 109.9,
-                                "cumulative": 209.8
-                            },
-                            {
-                                "interval": 109.95,
-                                "cumulative": 209.9
-                            },
-                            {
-                                "interval": 109.99,
-                                "cumulative": 209.98
-                            },
-                            {
-                                "interval": 110.0,
-                                "cumulative": 210.0
-                            }
-                        ]
-                    },
-                    {
-                        "name": "h2",
-                        "values": [
-                            {
-                                "interval": null,
-                                "cumulative": 100.0
-                            },
-                            {
-                                "interval": null,
-                                "cumulative": 102.5
-                            },
-                            {
-                                "interval": null,
-                                "cumulative": 105.0
-                            },
-                            {
-                                "interval": null,
-                                "cumulative": 107.5
-                            },
-                            {
-                                "interval": null,
-                                "cumulative": 109.0
-                            },
-                            {
-                                "interval": null,
-                                "cumulative": 109.5
-                            },
-                            {
-                                "interval": null,
-                                "cumulative": 109.9
-                            },
-                            {
-                                "interval": null,
-                                "cumulative": 109.95
-                            },
-                            {
-                                "interval": null,
-                                "cumulative": 109.99
-                            },
-                            {
-                                "interval": null,
-                                "cumulative": 110.0
-                            }
-                        ]
-                    }
-                ]
-            }
-        }
-    ]
-})EOF";
-
-  EXPECT_THAT(expected_json, JsonStringEq(actual_json));
-  store_->shutdownThreading();
-}
-
-TEST_P(AdminStatsTest, UsedOnlyStatsAsJson) {
-  InSequence s;
-  store_->initializeThreading(main_thread_dispatcher_, tls_);
-
-  Stats::Histogram& h1 = store_->histogramFromString("h1", Stats::Histogram::Unit::Unspecified);
-  Stats::Histogram& h2 = store_->histogramFromString("h2", Stats::Histogram::Unit::Unspecified);
-
-  EXPECT_EQ("h1", h1.name());
-  EXPECT_EQ("h2", h2.name());
-
-  EXPECT_CALL(sink_, onHistogramComplete(Ref(h1), 200));
-  h1.recordValue(200);
-
-  store_->mergeHistograms([]() -> void {});
-
-  // Again record a new value in h1 so that it has both interval and cumulative values.
-  // h2 should only have cumulative values.
-  EXPECT_CALL(sink_, onHistogramComplete(Ref(h1), 100));
-  h1.recordValue(100);
-
-  store_->mergeHistograms([]() -> void {});
-
-  std::map<std::string, uint64_t> all_stats;
-  std::map<std::string, std::string> all_text_readouts;
-  std::string actual_json =
-      statsAsJsonHandler(all_stats, all_text_readouts, store_->histograms(), true);
-
-  // Expected JSON should not have h2 values as it is not used.
-  const std::string expected_json = R"EOF({
-    "stats": [
-        {
-            "histograms": {
-                "supported_quantiles": [
-                    0.0,
-                    25.0,
-                    50.0,
-                    75.0,
-                    90.0,
-                    95.0,
-                    99.0,
-                    99.5,
-                    99.9,
-                    100.0
-                ],
-                "computed_quantiles": [
-                    {
-                        "name": "h1",
-                        "values": [
-                            {
-                                "interval": 100.0,
-                                "cumulative": 100.0
-                            },
-                            {
-                                "interval": 102.5,
-                                "cumulative": 105.0
-                            },
-                            {
-                                "interval": 105.0,
-                                "cumulative": 110.0
-                            },
-                            {
-                                "interval": 107.5,
-                                "cumulative": 205.0
-                            },
-                            {
-                                "interval": 109.0,
-                                "cumulative": 208.0
-                            },
-                            {
-                                "interval": 109.5,
-                                "cumulative": 209.0
-                            },
-                            {
-                                "interval": 109.9,
-                                "cumulative": 209.8
-                            },
-                            {
-                                "interval": 109.95,
-                                "cumulative": 209.9
-                            },
-                            {
-                                "interval": 109.99,
-                                "cumulative": 209.98
-                            },
-                            {
-                                "interval": 110.0,
-                                "cumulative": 210.0
-                            }
-                        ]
-                    }
-                ]
-            }
-        }
-    ]
-})EOF";
-
-  EXPECT_THAT(expected_json, JsonStringEq(actual_json));
-  store_->shutdownThreading();
-}
-
-TEST_P(AdminStatsTest, StatsAsJsonFilterString) {
-  InSequence s;
-  store_->initializeThreading(main_thread_dispatcher_, tls_);
-
-  Stats::Histogram& h1 = store_->histogramFromString("h1", Stats::Histogram::Unit::Unspecified);
-  Stats::Histogram& h2 = store_->histogramFromString("h2", Stats::Histogram::Unit::Unspecified);
-
-  EXPECT_CALL(sink_, onHistogramComplete(Ref(h1), 200));
-  h1.recordValue(200);
-
-  EXPECT_CALL(sink_, onHistogramComplete(Ref(h2), 100));
-  h2.recordValue(100);
-
-  store_->mergeHistograms([]() -> void {});
-
-  // Again record a new value in h1 so that it has both interval and cumulative values.
-  // h2 should only have cumulative values.
-  EXPECT_CALL(sink_, onHistogramComplete(Ref(h1), 100));
-  h1.recordValue(100);
-
-  store_->mergeHistograms([]() -> void {});
-
-  std::map<std::string, uint64_t> all_stats;
-  std::map<std::string, std::string> all_text_readouts;
-  std::string actual_json =
-      statsAsJsonHandler(all_stats, all_text_readouts, store_->histograms(), false,
-                         absl::optional<std::regex>{std::regex("[a-z]1")});
-
-  // Because this is a filter case, we don't expect to see any stats except for those containing
-  // "h1" in their name.
-  const std::string expected_json = R"EOF({
-    "stats": [
-        {
-            "histograms": {
-                "supported_quantiles": [
-                    0.0,
-                    25.0,
-                    50.0,
-                    75.0,
-                    90.0,
-                    95.0,
-                    99.0,
-                    99.5,
-                    99.9,
-                    100.0
-                ],
-                "computed_quantiles": [
-                    {
-                        "name": "h1",
-                        "values": [
-                            {
-                                "interval": 100.0,
-                                "cumulative": 100.0
-                            },
-                            {
-                                "interval": 102.5,
-                                "cumulative": 105.0
-                            },
-                            {
-                                "interval": 105.0,
-                                "cumulative": 110.0
-                            },
-                            {
-                                "interval": 107.5,
-                                "cumulative": 205.0
-                            },
-                            {
-                                "interval": 109.0,
-                                "cumulative": 208.0
-                            },
-                            {
-                                "interval": 109.5,
-                                "cumulative": 209.0
-                            },
-                            {
-                                "interval": 109.9,
-                                "cumulative": 209.8
-                            },
-                            {
-                                "interval": 109.95,
-                                "cumulative": 209.9
-                            },
-                            {
-                                "interval": 109.99,
-                                "cumulative": 209.98
-                            },
-                            {
-                                "interval": 110.0,
-                                "cumulative": 210.0
-                            }
-                        ]
-                    }
-                ]
-            }
-        }
-    ]
-})EOF";
-
-  EXPECT_THAT(expected_json, JsonStringEq(actual_json));
-  store_->shutdownThreading();
-}
-
-TEST_P(AdminStatsTest, UsedOnlyStatsAsJsonFilterString) {
-  InSequence s;
-  store_->initializeThreading(main_thread_dispatcher_, tls_);
-
-  Stats::Histogram& h1 = store_->histogramFromString(
-      "h1_matches", Stats::Histogram::Unit::Unspecified); // Will match, be used, and print
-  Stats::Histogram& h2 = store_->histogramFromString(
-      "h2_matches", Stats::Histogram::Unit::Unspecified); // Will match but not be used
-  Stats::Histogram& h3 = store_->histogramFromString(
-      "h3_not", Stats::Histogram::Unit::Unspecified); // Will be used but not match
-
-  EXPECT_EQ("h1_matches", h1.name());
-  EXPECT_EQ("h2_matches", h2.name());
-  EXPECT_EQ("h3_not", h3.name());
-
-  EXPECT_CALL(sink_, onHistogramComplete(Ref(h1), 200));
-  h1.recordValue(200);
-  EXPECT_CALL(sink_, onHistogramComplete(Ref(h3), 200));
-  h3.recordValue(200);
-
-  store_->mergeHistograms([]() -> void {});
-
-  // Again record a new value in h1 and h3 so that they have both interval and cumulative values.
-  // h2 should only have cumulative values.
-  EXPECT_CALL(sink_, onHistogramComplete(Ref(h1), 100));
-  h1.recordValue(100);
-  EXPECT_CALL(sink_, onHistogramComplete(Ref(h3), 100));
-  h3.recordValue(100);
-
-  store_->mergeHistograms([]() -> void {});
-
-  std::map<std::string, uint64_t> all_stats;
-  std::map<std::string, std::string> all_text_readouts;
-  std::string actual_json =
-      statsAsJsonHandler(all_stats, all_text_readouts, store_->histograms(), true,
-                         absl::optional<std::regex>{std::regex("h[12]")});
-
-  // Expected JSON should not have h2 values as it is not used, and should not have h3 values as
-  // they are used but do not match.
-  const std::string expected_json = R"EOF({
-    "stats": [
-        {
-            "histograms": {
-                "supported_quantiles": [
-                    0.0,
-                    25.0,
-                    50.0,
-                    75.0,
-                    90.0,
-                    95.0,
-                    99.0,
-                    99.5,
-                    99.9,
-                    100.0
-                ],
-                "computed_quantiles": [
-                    {
-                        "name": "h1_matches",
-                        "values": [
-                            {
-                                "interval": 100.0,
-                                "cumulative": 100.0
-                            },
-                            {
-                                "interval": 102.5,
-                                "cumulative": 105.0
-                            },
-                            {
-                                "interval": 105.0,
-                                "cumulative": 110.0
-                            },
-                            {
-                                "interval": 107.5,
-                                "cumulative": 205.0
-                            },
-                            {
-                                "interval": 109.0,
-                                "cumulative": 208.0
-                            },
-                            {
-                                "interval": 109.5,
-                                "cumulative": 209.0
-                            },
-                            {
-                                "interval": 109.9,
-                                "cumulative": 209.8
-                            },
-                            {
-                                "interval": 109.95,
-                                "cumulative": 209.9
-                            },
-                            {
-                                "interval": 109.99,
-                                "cumulative": 209.98
-                            },
-                            {
-                                "interval": 110.0,
-                                "cumulative": 210.0
-                            }
-                        ]
-                    }
-                ]
-            }
-        }
-    ]
-})EOF";
-
-  EXPECT_THAT(expected_json, JsonStringEq(actual_json));
-  store_->shutdownThreading();
-}
-
-class AdminInstanceTest : public testing::TestWithParam<Network::Address::IpVersion> {
-public:
-  AdminInstanceTest()
-      : address_out_path_(TestEnvironment::temporaryPath("admin.address")),
-        cpu_profile_path_(TestEnvironment::temporaryPath("envoy.prof")),
-        admin_(cpu_profile_path_, server_), request_headers_{{":path", "/"}},
-        admin_filter_(admin_.createCallbackFunction()) {
-    admin_.startHttpListener("/dev/null", address_out_path_,
-                             Network::Test::getCanonicalLoopbackAddress(GetParam()), nullptr,
-                             listener_scope_.createScope("listener.admin."));
-    EXPECT_EQ(std::chrono::milliseconds(100), admin_.drainTimeout());
-    admin_.tracingStats().random_sampling_.inc();
-    EXPECT_TRUE(admin_.setCurrentClientCertDetails().empty());
-    admin_filter_.setDecoderFilterCallbacks(callbacks_);
-  }
-
-  Http::Code runCallback(absl::string_view path_and_query,
-                         Http::ResponseHeaderMap& response_headers, Buffer::Instance& response,
-                         absl::string_view method, absl::string_view body = absl::string_view()) {
-    if (!body.empty()) {
-      request_headers_.setReferenceContentType(
-          Http::Headers::get().ContentTypeValues.FormUrlEncoded);
-      callbacks_.buffer_ = std::make_unique<Buffer::OwnedImpl>(body);
-    }
-
-    request_headers_.setMethod(method);
-    admin_filter_.decodeHeaders(request_headers_, false);
-
-    return admin_.runCallback(path_and_query, response_headers, response, admin_filter_);
-  }
-
-  Http::Code getCallback(absl::string_view path_and_query,
-                         Http::ResponseHeaderMap& response_headers, Buffer::Instance& response) {
-    return runCallback(path_and_query, response_headers, response,
-                       Http::Headers::get().MethodValues.Get);
-  }
-
-  Http::Code postCallback(absl::string_view path_and_query,
-                          Http::ResponseHeaderMap& response_headers, Buffer::Instance& response) {
-    return runCallback(path_and_query, response_headers, response,
-                       Http::Headers::get().MethodValues.Post);
-  }
-
-  std::string address_out_path_;
-  std::string cpu_profile_path_;
-  NiceMock<MockInstance> server_;
-  Stats::IsolatedStoreImpl listener_scope_;
-  AdminImpl admin_;
-  Http::TestRequestHeaderMapImpl request_headers_;
-  Server::AdminFilter admin_filter_;
-  NiceMock<Http::MockStreamDecoderFilterCallbacks> callbacks_;
-};
 
 INSTANTIATE_TEST_SUITE_P(IpVersions, AdminInstanceTest,
                          testing::ValuesIn(TestEnvironment::getIpVersionsForTest()),
@@ -680,37 +125,6 @@ TEST_P(AdminInstanceTest, AdminBadProfiler) {
                            admin_bad_profile_path.runCallback("/cpuprofiler?enable=y", header_map,
                                                               data, admin_filter_)));
   EXPECT_FALSE(Profiler::Cpu::profilerEnabled());
-}
-
-TEST_P(AdminInstanceTest, StatsInvalidRegex) {
-  Http::ResponseHeaderMapImpl header_map;
-  Buffer::OwnedImpl data;
-  EXPECT_LOG_CONTAINS(
-      "error", "Invalid regex: ",
-      EXPECT_EQ(Http::Code::BadRequest, getCallback("/stats?filter=*.test", header_map, data)));
-
-  // Note: depending on the library, the detailed error message might be one of:
-  //   "One of *?+{ was not preceded by a valid regular expression."
-  //   "regex_error"
-  // but we always precede by 'Invalid regex: "'.
-  EXPECT_THAT(data.toString(), StartsWith("Invalid regex: \""));
-  EXPECT_THAT(data.toString(), EndsWith("\"\n"));
-}
-
-TEST_P(AdminInstanceTest, PrometheusStatsInvalidRegex) {
-  Http::ResponseHeaderMapImpl header_map;
-  Buffer::OwnedImpl data;
-  EXPECT_LOG_CONTAINS(
-      "error", ": *.ptest",
-      EXPECT_EQ(Http::Code::BadRequest,
-                getCallback("/stats?format=prometheus&filter=*.ptest", header_map, data)));
-
-  // Note: depending on the library, the detailed error message might be one of:
-  //   "One of *?+{ was not preceded by a valid regular expression."
-  //   "regex_error"
-  // but we always precede by 'Invalid regex: "'.
-  EXPECT_THAT(data.toString(), StartsWith("Invalid regex: \""));
-  EXPECT_THAT(data.toString(), EndsWith("\"\n"));
 }
 
 TEST_P(AdminInstanceTest, WriteAddressToFile) {
@@ -1202,13 +616,6 @@ TEST_P(AdminInstanceTest, ReopenLogs) {
   EXPECT_EQ(Http::Code::OK, postCallback("/reopen_logs", header_map, response));
 }
 
-TEST_P(AdminInstanceTest, TracingStatsDisabled) {
-  const std::string& name = admin_.tracingStats().service_forced_.name();
-  for (const Stats::CounterSharedPtr& counter : server_.stats().counters()) {
-    EXPECT_NE(counter->name(), name) << "Unexpected tracing stat found in server stats: " << name;
-  }
-}
-
 TEST_P(AdminInstanceTest, ClustersJson) {
   Upstream::ClusterManager::ClusterInfoMap cluster_map;
   ON_CALL(server_.cluster_manager_, clusters()).WillByDefault(ReturnPointee(&cluster_map));
@@ -1518,29 +925,6 @@ TEST_P(AdminInstanceTest, GetReadyRequest) {
               HasSubstr("text/plain"));
 }
 
-TEST_P(AdminInstanceTest, GetRequestJson) {
-  Http::ResponseHeaderMapImpl response_headers;
-  std::string body;
-  EXPECT_EQ(Http::Code::OK, admin_.request("/stats?format=json", "GET", response_headers, body));
-  EXPECT_THAT(body, HasSubstr("{\"stats\":["));
-  EXPECT_THAT(std::string(response_headers.ContentType()->value().getStringView()),
-              HasSubstr("application/json"));
-}
-
-TEST_P(AdminInstanceTest, RecentLookups) {
-  Http::ResponseHeaderMapImpl response_headers;
-  std::string body;
-
-  // Recent lookup tracking is disabled by default.
-  EXPECT_EQ(Http::Code::OK, admin_.request("/stats/recentlookups", "GET", response_headers, body));
-  EXPECT_THAT(body, HasSubstr("Lookup tracking is not enabled"));
-  EXPECT_THAT(std::string(response_headers.ContentType()->value().getStringView()),
-              HasSubstr("text/plain"));
-
-  // We can't test RecentLookups in admin unit tests as it doesn't work with a
-  // fake symbol table. However we cover this solidly in integration tests.
-}
-
 TEST_P(AdminInstanceTest, PostRequest) {
   Http::ResponseHeaderMapImpl response_headers;
   std::string body;
@@ -1549,443 +933,6 @@ TEST_P(AdminInstanceTest, PostRequest) {
   EXPECT_EQ(body, "OK\n");
   EXPECT_THAT(std::string(response_headers.ContentType()->value().getStringView()),
               HasSubstr("text/plain"));
-}
-
-class HistogramWrapper {
-public:
-  HistogramWrapper() : histogram_(hist_alloc()) {}
-
-  ~HistogramWrapper() { hist_free(histogram_); }
-
-  const histogram_t* getHistogram() { return histogram_; }
-
-  void setHistogramValues(const std::vector<uint64_t>& values) {
-    for (uint64_t value : values) {
-      hist_insert_intscale(histogram_, value, 0, 1);
-    }
-  }
-
-  void setHistogramValuesWithCounts(const std::vector<std::pair<uint64_t, uint64_t>>& values) {
-    for (std::pair<uint64_t, uint64_t> cv : values) {
-      hist_insert_intscale(histogram_, cv.first, 0, cv.second);
-    }
-  }
-
-private:
-  histogram_t* histogram_;
-};
-
-class PrometheusStatsFormatterTest : public testing::Test {
-protected:
-  PrometheusStatsFormatterTest()
-      : symbol_table_(Stats::SymbolTableCreator::makeSymbolTable()), alloc_(*symbol_table_),
-        pool_(*symbol_table_) {}
-
-  ~PrometheusStatsFormatterTest() override { clearStorage(); }
-
-  void addCounter(const std::string& name, Stats::StatNameTagVector cluster_tags) {
-    Stats::StatNameManagedStorage storage(name, *symbol_table_);
-    Stats::StatName stat_name = storage.statName();
-    counters_.push_back(alloc_.makeCounter(stat_name, stat_name, cluster_tags));
-  }
-
-  void addGauge(const std::string& name, Stats::StatNameTagVector cluster_tags) {
-    Stats::StatNameManagedStorage storage(name, *symbol_table_);
-    Stats::StatName stat_name = storage.statName();
-    gauges_.push_back(
-        alloc_.makeGauge(stat_name, stat_name, cluster_tags, Stats::Gauge::ImportMode::Accumulate));
-  }
-
-  void addHistogram(const Stats::ParentHistogramSharedPtr histogram) {
-    histograms_.push_back(histogram);
-  }
-
-  using MockHistogramSharedPtr = Stats::RefcountPtr<NiceMock<Stats::MockParentHistogram>>;
-  MockHistogramSharedPtr makeHistogram() {
-    return MockHistogramSharedPtr(new NiceMock<Stats::MockParentHistogram>());
-  }
-
-  Stats::StatName makeStat(absl::string_view name) { return pool_.add(name); }
-
-  void clearStorage() {
-    pool_.clear();
-    counters_.clear();
-    gauges_.clear();
-    histograms_.clear();
-    EXPECT_EQ(0, symbol_table_->numSymbols());
-  }
-
-  Stats::SymbolTablePtr symbol_table_;
-  Stats::AllocatorImpl alloc_;
-  Stats::StatNamePool pool_;
-  std::vector<Stats::CounterSharedPtr> counters_;
-  std::vector<Stats::GaugeSharedPtr> gauges_;
-  std::vector<Stats::ParentHistogramSharedPtr> histograms_;
-};
-
-TEST_F(PrometheusStatsFormatterTest, MetricName) {
-  std::string raw = "vulture.eats-liver";
-  std::string expected = "envoy_vulture_eats_liver";
-  auto actual = PrometheusStatsFormatter::metricName(raw);
-  EXPECT_EQ(expected, actual);
-}
-
-TEST_F(PrometheusStatsFormatterTest, SanitizeMetricName) {
-  std::string raw = "An.artist.plays-violin@019street";
-  std::string expected = "envoy_An_artist_plays_violin_019street";
-  auto actual = PrometheusStatsFormatter::metricName(raw);
-  EXPECT_EQ(expected, actual);
-}
-
-TEST_F(PrometheusStatsFormatterTest, SanitizeMetricNameDigitFirst) {
-  std::string raw = "3.artists.play-violin@019street";
-  std::string expected = "envoy_3_artists_play_violin_019street";
-  auto actual = PrometheusStatsFormatter::metricName(raw);
-  EXPECT_EQ(expected, actual);
-}
-
-TEST_F(PrometheusStatsFormatterTest, FormattedTags) {
-  std::vector<Stats::Tag> tags;
-  Stats::Tag tag1 = {"a.tag-name", "a.tag-value"};
-  Stats::Tag tag2 = {"another_tag_name", "another_tag-value"};
-  tags.push_back(tag1);
-  tags.push_back(tag2);
-  std::string expected = "a_tag_name=\"a.tag-value\",another_tag_name=\"another_tag-value\"";
-  auto actual = PrometheusStatsFormatter::formattedTags(tags);
-  EXPECT_EQ(expected, actual);
-}
-
-TEST_F(PrometheusStatsFormatterTest, MetricNameCollison) {
-
-  // Create two counters and two gauges with each pair having the same name,
-  // but having different tag names and values.
-  //`statsAsPrometheus()` should return two implying it found two unique stat names
-
-  addCounter("cluster.test_cluster_1.upstream_cx_total",
-             {{makeStat("a.tag-name"), makeStat("a.tag-value")}});
-  addCounter("cluster.test_cluster_1.upstream_cx_total",
-             {{makeStat("another_tag_name"), makeStat("another_tag-value")}});
-  addGauge("cluster.test_cluster_2.upstream_cx_total",
-           {{makeStat("another_tag_name_3"), makeStat("another_tag_3-value")}});
-  addGauge("cluster.test_cluster_2.upstream_cx_total",
-           {{makeStat("another_tag_name_4"), makeStat("another_tag_4-value")}});
-
-  Buffer::OwnedImpl response;
-  auto size = PrometheusStatsFormatter::statsAsPrometheus(counters_, gauges_, histograms_, response,
-                                                          false, absl::nullopt);
-  EXPECT_EQ(2UL, size);
-}
-
-TEST_F(PrometheusStatsFormatterTest, UniqueMetricName) {
-
-  // Create two counters and two gauges, all with unique names.
-  // statsAsPrometheus() should return four implying it found
-  // four unique stat names.
-
-  addCounter("cluster.test_cluster_1.upstream_cx_total",
-             {{makeStat("a.tag-name"), makeStat("a.tag-value")}});
-  addCounter("cluster.test_cluster_2.upstream_cx_total",
-             {{makeStat("another_tag_name"), makeStat("another_tag-value")}});
-  addGauge("cluster.test_cluster_3.upstream_cx_total",
-           {{makeStat("another_tag_name_3"), makeStat("another_tag_3-value")}});
-  addGauge("cluster.test_cluster_4.upstream_cx_total",
-           {{makeStat("another_tag_name_4"), makeStat("another_tag_4-value")}});
-
-  Buffer::OwnedImpl response;
-  auto size = PrometheusStatsFormatter::statsAsPrometheus(counters_, gauges_, histograms_, response,
-                                                          false, absl::nullopt);
-  EXPECT_EQ(4UL, size);
-}
-
-TEST_F(PrometheusStatsFormatterTest, HistogramWithNoValuesAndNoTags) {
-  HistogramWrapper h1_cumulative;
-  h1_cumulative.setHistogramValues(std::vector<uint64_t>(0));
-  Stats::HistogramStatisticsImpl h1_cumulative_statistics(h1_cumulative.getHistogram());
-
-  auto histogram = makeHistogram();
-  histogram->name_ = "histogram1";
-  histogram->used_ = true;
-  ON_CALL(*histogram, cumulativeStatistics())
-      .WillByDefault(testing::ReturnRef(h1_cumulative_statistics));
-
-  addHistogram(histogram);
-
-  Buffer::OwnedImpl response;
-  auto size = PrometheusStatsFormatter::statsAsPrometheus(counters_, gauges_, histograms_, response,
-                                                          false, absl::nullopt);
-  EXPECT_EQ(1UL, size);
-
-  const std::string expected_output = R"EOF(# TYPE envoy_histogram1 histogram
-envoy_histogram1_bucket{le="0.5"} 0
-envoy_histogram1_bucket{le="1"} 0
-envoy_histogram1_bucket{le="5"} 0
-envoy_histogram1_bucket{le="10"} 0
-envoy_histogram1_bucket{le="25"} 0
-envoy_histogram1_bucket{le="50"} 0
-envoy_histogram1_bucket{le="100"} 0
-envoy_histogram1_bucket{le="250"} 0
-envoy_histogram1_bucket{le="500"} 0
-envoy_histogram1_bucket{le="1000"} 0
-envoy_histogram1_bucket{le="2500"} 0
-envoy_histogram1_bucket{le="5000"} 0
-envoy_histogram1_bucket{le="10000"} 0
-envoy_histogram1_bucket{le="30000"} 0
-envoy_histogram1_bucket{le="60000"} 0
-envoy_histogram1_bucket{le="300000"} 0
-envoy_histogram1_bucket{le="600000"} 0
-envoy_histogram1_bucket{le="1800000"} 0
-envoy_histogram1_bucket{le="3600000"} 0
-envoy_histogram1_bucket{le="+Inf"} 0
-envoy_histogram1_sum{} 0
-envoy_histogram1_count{} 0
-)EOF";
-
-  EXPECT_EQ(expected_output, response.toString());
-}
-
-TEST_F(PrometheusStatsFormatterTest, HistogramWithHighCounts) {
-  HistogramWrapper h1_cumulative;
-
-  // Force large counts to prove that the +Inf bucket doesn't overflow to scientific notation.
-  h1_cumulative.setHistogramValuesWithCounts(std::vector<std::pair<uint64_t, uint64_t>>({
-      {1, 100000},
-      {100, 1000000},
-      {1000, 100000000},
-  }));
-
-  Stats::HistogramStatisticsImpl h1_cumulative_statistics(h1_cumulative.getHistogram());
-
-  auto histogram = makeHistogram();
-  histogram->name_ = "histogram1";
-  histogram->used_ = true;
-  ON_CALL(*histogram, cumulativeStatistics())
-      .WillByDefault(testing::ReturnRef(h1_cumulative_statistics));
-
-  addHistogram(histogram);
-
-  Buffer::OwnedImpl response;
-  auto size = PrometheusStatsFormatter::statsAsPrometheus(counters_, gauges_, histograms_, response,
-                                                          false, absl::nullopt);
-  EXPECT_EQ(1UL, size);
-
-  const std::string expected_output = R"EOF(# TYPE envoy_histogram1 histogram
-envoy_histogram1_bucket{le="0.5"} 0
-envoy_histogram1_bucket{le="1"} 0
-envoy_histogram1_bucket{le="5"} 100000
-envoy_histogram1_bucket{le="10"} 100000
-envoy_histogram1_bucket{le="25"} 100000
-envoy_histogram1_bucket{le="50"} 100000
-envoy_histogram1_bucket{le="100"} 100000
-envoy_histogram1_bucket{le="250"} 1100000
-envoy_histogram1_bucket{le="500"} 1100000
-envoy_histogram1_bucket{le="1000"} 1100000
-envoy_histogram1_bucket{le="2500"} 101100000
-envoy_histogram1_bucket{le="5000"} 101100000
-envoy_histogram1_bucket{le="10000"} 101100000
-envoy_histogram1_bucket{le="30000"} 101100000
-envoy_histogram1_bucket{le="60000"} 101100000
-envoy_histogram1_bucket{le="300000"} 101100000
-envoy_histogram1_bucket{le="600000"} 101100000
-envoy_histogram1_bucket{le="1800000"} 101100000
-envoy_histogram1_bucket{le="3600000"} 101100000
-envoy_histogram1_bucket{le="+Inf"} 101100000
-envoy_histogram1_sum{} 105105105000
-envoy_histogram1_count{} 101100000
-)EOF";
-
-  EXPECT_EQ(expected_output, response.toString());
-}
-
-TEST_F(PrometheusStatsFormatterTest, OutputWithAllMetricTypes) {
-  addCounter("cluster.test_1.upstream_cx_total",
-             {{makeStat("a.tag-name"), makeStat("a.tag-value")}});
-  addCounter("cluster.test_2.upstream_cx_total",
-             {{makeStat("another_tag_name"), makeStat("another_tag-value")}});
-  addGauge("cluster.test_3.upstream_cx_total",
-           {{makeStat("another_tag_name_3"), makeStat("another_tag_3-value")}});
-  addGauge("cluster.test_4.upstream_cx_total",
-           {{makeStat("another_tag_name_4"), makeStat("another_tag_4-value")}});
-
-  const std::vector<uint64_t> h1_values = {50, 20, 30, 70, 100, 5000, 200};
-  HistogramWrapper h1_cumulative;
-  h1_cumulative.setHistogramValues(h1_values);
-  Stats::HistogramStatisticsImpl h1_cumulative_statistics(h1_cumulative.getHistogram());
-
-  auto histogram1 = makeHistogram();
-  histogram1->name_ = "cluster.test_1.upstream_rq_time";
-  histogram1->unit_ = Stats::Histogram::Unit::Milliseconds;
-  histogram1->used_ = true;
-  histogram1->setTags({Stats::Tag{"key1", "value1"}, Stats::Tag{"key2", "value2"}});
-  addHistogram(histogram1);
-  EXPECT_CALL(*histogram1, cumulativeStatistics())
-      .WillOnce(testing::ReturnRef(h1_cumulative_statistics));
-
-  Buffer::OwnedImpl response;
-  auto size = PrometheusStatsFormatter::statsAsPrometheus(counters_, gauges_, histograms_, response,
-                                                          false, absl::nullopt);
-  EXPECT_EQ(5UL, size);
-
-  const std::string expected_output = R"EOF(# TYPE envoy_cluster_test_1_upstream_cx_total counter
-envoy_cluster_test_1_upstream_cx_total{a_tag_name="a.tag-value"} 0
-# TYPE envoy_cluster_test_2_upstream_cx_total counter
-envoy_cluster_test_2_upstream_cx_total{another_tag_name="another_tag-value"} 0
-# TYPE envoy_cluster_test_3_upstream_cx_total gauge
-envoy_cluster_test_3_upstream_cx_total{another_tag_name_3="another_tag_3-value"} 0
-# TYPE envoy_cluster_test_4_upstream_cx_total gauge
-envoy_cluster_test_4_upstream_cx_total{another_tag_name_4="another_tag_4-value"} 0
-# TYPE envoy_cluster_test_1_upstream_rq_time histogram
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="0.5"} 0
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="1"} 0
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="5"} 0
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="10"} 0
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="25"} 1
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="50"} 2
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="100"} 4
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="250"} 6
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="500"} 6
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="1000"} 6
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="2500"} 6
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="5000"} 6
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="10000"} 7
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="30000"} 7
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="60000"} 7
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="300000"} 7
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="600000"} 7
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="1800000"} 7
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="3600000"} 7
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="+Inf"} 7
-envoy_cluster_test_1_upstream_rq_time_sum{key1="value1",key2="value2"} 5532
-envoy_cluster_test_1_upstream_rq_time_count{key1="value1",key2="value2"} 7
-)EOF";
-
-  EXPECT_EQ(expected_output, response.toString());
-}
-
-TEST_F(PrometheusStatsFormatterTest, OutputWithUsedOnly) {
-  addCounter("cluster.test_1.upstream_cx_total",
-             {{makeStat("a.tag-name"), makeStat("a.tag-value")}});
-  addCounter("cluster.test_2.upstream_cx_total",
-             {{makeStat("another_tag_name"), makeStat("another_tag-value")}});
-  addGauge("cluster.test_3.upstream_cx_total",
-           {{makeStat("another_tag_name_3"), makeStat("another_tag_3-value")}});
-  addGauge("cluster.test_4.upstream_cx_total",
-           {{makeStat("another_tag_name_4"), makeStat("another_tag_4-value")}});
-
-  const std::vector<uint64_t> h1_values = {50, 20, 30, 70, 100, 5000, 200};
-  HistogramWrapper h1_cumulative;
-  h1_cumulative.setHistogramValues(h1_values);
-  Stats::HistogramStatisticsImpl h1_cumulative_statistics(h1_cumulative.getHistogram());
-
-  auto histogram1 = makeHistogram();
-  histogram1->name_ = "cluster.test_1.upstream_rq_time";
-  histogram1->unit_ = Stats::Histogram::Unit::Milliseconds;
-  histogram1->used_ = true;
-  histogram1->setTags({Stats::Tag{"key1", "value1"}, Stats::Tag{"key2", "value2"}});
-  addHistogram(histogram1);
-  EXPECT_CALL(*histogram1, cumulativeStatistics())
-      .WillOnce(testing::ReturnRef(h1_cumulative_statistics));
-
-  Buffer::OwnedImpl response;
-  auto size = PrometheusStatsFormatter::statsAsPrometheus(counters_, gauges_, histograms_, response,
-                                                          true, absl::nullopt);
-  EXPECT_EQ(1UL, size);
-
-  const std::string expected_output = R"EOF(# TYPE envoy_cluster_test_1_upstream_rq_time histogram
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="0.5"} 0
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="1"} 0
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="5"} 0
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="10"} 0
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="25"} 1
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="50"} 2
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="100"} 4
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="250"} 6
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="500"} 6
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="1000"} 6
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="2500"} 6
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="5000"} 6
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="10000"} 7
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="30000"} 7
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="60000"} 7
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="300000"} 7
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="600000"} 7
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="1800000"} 7
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="3600000"} 7
-envoy_cluster_test_1_upstream_rq_time_bucket{key1="value1",key2="value2",le="+Inf"} 7
-envoy_cluster_test_1_upstream_rq_time_sum{key1="value1",key2="value2"} 5532
-envoy_cluster_test_1_upstream_rq_time_count{key1="value1",key2="value2"} 7
-)EOF";
-
-  EXPECT_EQ(expected_output, response.toString());
-}
-
-TEST_F(PrometheusStatsFormatterTest, OutputWithUsedOnlyHistogram) {
-  const std::vector<uint64_t> h1_values = {};
-  HistogramWrapper h1_cumulative;
-  h1_cumulative.setHistogramValues(h1_values);
-  Stats::HistogramStatisticsImpl h1_cumulative_statistics(h1_cumulative.getHistogram());
-
-  auto histogram1 = makeHistogram();
-  histogram1->name_ = "cluster.test_1.upstream_rq_time";
-  histogram1->unit_ = Stats::Histogram::Unit::Milliseconds;
-  histogram1->used_ = false;
-  histogram1->setTags({Stats::Tag{"key1", "value1"}, Stats::Tag{"key2", "value2"}});
-  addHistogram(histogram1);
-
-  {
-    const bool used_only = true;
-    EXPECT_CALL(*histogram1, cumulativeStatistics()).Times(0);
-
-    Buffer::OwnedImpl response;
-    auto size = PrometheusStatsFormatter::statsAsPrometheus(counters_, gauges_, histograms_,
-                                                            response, used_only, absl::nullopt);
-    EXPECT_EQ(0UL, size);
-  }
-
-  {
-    const bool used_only = false;
-    EXPECT_CALL(*histogram1, cumulativeStatistics())
-        .WillOnce(testing::ReturnRef(h1_cumulative_statistics));
-
-    Buffer::OwnedImpl response;
-    auto size = PrometheusStatsFormatter::statsAsPrometheus(counters_, gauges_, histograms_,
-                                                            response, used_only, absl::nullopt);
-    EXPECT_EQ(1UL, size);
-  }
-}
-
-TEST_F(PrometheusStatsFormatterTest, OutputWithRegexp) {
-  addCounter("cluster.test_1.upstream_cx_total",
-             {{makeStat("a.tag-name"), makeStat("a.tag-value")}});
-  addCounter("cluster.test_2.upstream_cx_total",
-             {{makeStat("another_tag_name"), makeStat("another_tag-value")}});
-  addGauge("cluster.test_3.upstream_cx_total",
-           {{makeStat("another_tag_name_3"), makeStat("another_tag_3-value")}});
-  addGauge("cluster.test_4.upstream_cx_total",
-           {{makeStat("another_tag_name_4"), makeStat("another_tag_4-value")}});
-
-  const std::vector<uint64_t> h1_values = {50, 20, 30, 70, 100, 5000, 200};
-  HistogramWrapper h1_cumulative;
-  h1_cumulative.setHistogramValues(h1_values);
-  Stats::HistogramStatisticsImpl h1_cumulative_statistics(h1_cumulative.getHistogram());
-
-  auto histogram1 = makeHistogram();
-  histogram1->name_ = "cluster.test_1.upstream_rq_time";
-  histogram1->unit_ = Stats::Histogram::Unit::Milliseconds;
-  histogram1->setTags({Stats::Tag{"key1", "value1"}, Stats::Tag{"key2", "value2"}});
-  addHistogram(histogram1);
-
-  Buffer::OwnedImpl response;
-  auto size = PrometheusStatsFormatter::statsAsPrometheus(
-      counters_, gauges_, histograms_, response, false,
-      absl::optional<std::regex>{std::regex("cluster.test_1.upstream_cx_total")});
-  EXPECT_EQ(1UL, size);
-
-  const std::string expected_output =
-      R"EOF(# TYPE envoy_cluster_test_1_upstream_cx_total counter
-envoy_cluster_test_1_upstream_cx_total{a_tag_name="a.tag-value"} 0
-)EOF";
-
-  EXPECT_EQ(expected_output, response.toString());
 }
 
 } // namespace Server
